#832 Consolidation Part I

See merge request leos/core!716

+++ tools/repository/server/src/main/java/eu/europa/ec/leos/repository/repositories/DocumentMilestoneRepository.java
@@ -16,6 +16,8 @@ package eu.europa.ec.leos.repository.repositories;
 import eu.europa.ec.leos.repository.entities.DocumentMilestone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 
-public interface DocumentMilestoneRepository extends JpaRepository<DocumentMilestone, Long> {
+
+public interface DocumentMilestoneRepository extends JpaRepository<DocumentMilestone, BigDecimal> {
 }
